move functions over to seperate file;
