--- packages/data-mate/src/adapters/function-adapter/validators/field.ts
@@ -1,453 +1,0 @@
-import {
-    isObjectEntity,
-    get,
-    set,
-    isNotNil,
-    cloneDeep,
-    isNil,
-    unset,
-    isFunction
-} from '@terascope/utils';
-import {
-    FieldValidateConfig, InitialFunctionContext, FunctionContext,
-} from '../../../function-configs/interfaces';
-import { callValue } from '../utils';
-import { validateFunctionArgs } from '../../argument-validator';
-import { FunctionAdapterContext, DynamicFunctionAdapterContext } from '../interfaces';
-
-export function fieldValidationColumnExecution<T extends Record<string, any>>(
-    fnDef: FieldValidateConfig<T>,
-    configs: InitialFunctionContext<T>,
-) {
-    return function _fieldValidationColumnExecution(
-        input: unknown[],
-    ): (unknown|null)[] {
-        if (!Array.isArray(input)) {
-            throw new Error('Invalid input, expected an array of values');
-        }
-
-        if (isFunction(configs.args)) {
-            const dynamicContext = {
-                ...configs,
-                parent: input,
-            } as DynamicFunctionAdapterContext<T>;
-
-            return _fieldValidationDynamic<T>(fnDef, dynamicContext);
-        }
-
-        const functionContext: FunctionAdapterContext<T> = {
-            ...configs,
-            args: configs.args as T,
-            parent: input
-        };
-
-        return _fieldValidation<T>(fnDef, functionContext);
-    };
-}
-
-function _fieldValidationDynamic<T>(
-    fnDef: FieldValidateConfig<T>,
-    configs: DynamicFunctionAdapterContext<T>,
-) {
-    const results: (unknown | null)[] = [];
-    const input = configs.parent;
-
-    for (let i = 0; i < input.length; i++) {
-        const newArgs = configs.args(i, configs.parent) as T;
-        validateFunctionArgs(fnDef, newArgs);
-
-        const fnConfig: FunctionContext<T> = {
-            ...configs,
-            args: newArgs as T,
-        };
-
-        const fn = fnDef.create(fnConfig);
-        const value = input[i];
-
-        _processValidationValue(fn, value, results, configs.preserveNulls, i);
-    }
-
-    return results;
-}
-
-function _fieldValidation<T>(
-    fnDef: FieldValidateConfig<T>,
-    configs: FunctionAdapterContext<T>,
-) {
-    const fn = fnDef.create(configs);
-    const input = configs.parent as unknown[];
-
-    const results: (unknown | null)[] = [];
-
-    for (let i = 0; i < input.length; i++) {
-        const value = input[i];
-        _processValidationValue(fn, value, results, configs.preserveNulls, i);
-    }
-
-    return results;
-}
-
-function _processValidationValue(
-    fn: (input: unknown, index: number) => unknown,
-    value: unknown,
-    results: unknown[],
-    preserveNulls: boolean,
-    i: number
-) {
-    if (Array.isArray(value)) {
-        const fieldList = callValue(fn, value, preserveNulls, true, i);
-
-        if (fieldList.length > 0) {
-            results.push(fieldList);
-        } else if (preserveNulls) {
-            results.push(null);
-        }
-    } else if (isNotNil(value) && fn(value, i)) {
-        results.push(...callValue(fn, value, preserveNulls, true, i));
-    } else if (preserveNulls) {
-        results.push(null);
-    }
-}
-
-export function wholeFieldValidationColumnExecution<T extends Record<string, any>>(
-    fnDef: FieldValidateConfig<T>,
-    configs: InitialFunctionContext<T>,
-) {
-    return function _fieldValidationColumnExecution(
-        input: unknown[],
-    ): unknown[] {
-        const results: unknown[] = [];
-
-        if (isFunction(configs.args)) {
-            const dynamicContext = {
-                ...configs,
-                parent: input,
-            } as DynamicFunctionAdapterContext<T>;
-
-            input.forEach((value: unknown, index: number) => {
-                const newArgs = dynamicContext.args(index, input) as T;
-                validateFunctionArgs(fnDef, newArgs);
-
-                const fnConfig: FunctionContext<T> = {
-                    ...dynamicContext,
-                    args: newArgs as T,
-                };
-
-                const fn = fnDef.create(fnConfig);
-                const bool = fn(value, index);
-
-                if (bool) {
-                    results.push(value);
-                } else if (configs.preserveNulls) {
-                    results.push(null);
-                }
-            });
-
-            return results;
-        }
-
-        const functionContext: FunctionAdapterContext<T> = {
-            ...configs,
-            args: configs.args as T,
-            parent: input
-        };
-
-        const fn = fnDef.create(functionContext);
-
-        input.forEach((value: unknown, index: number) => {
-            const bool = fn(value, index);
-            if (bool) {
-                results.push(value);
-            } else if (configs.preserveNulls) {
-                results.push(null);
-            }
-        });
-
-        return results;
-    };
-}
-
-export function wholeFieldValidationRowExecution<T extends Record<string, any>>(
-    fnDef: FieldValidateConfig<T>,
-    configs: InitialFunctionContext<T>,
-): (input: Record<string, unknown>[]) => Record<string, unknown>[] {
-    return function _wholeFieldValidationRowExecution(
-        input: Record<string, unknown>[],
-    ): Record<string, unknown>[] {
-        if (isNil(configs.field)) throw new Error('Must provide a field option when running a row');
-
-        if (!Array.isArray(input)) {
-            throw new Error('Invalid input, expected an array of objects');
-        }
-
-        if (isFunction(configs.args)) {
-            const dynamicContext = {
-                ...configs,
-                parent: input,
-            } as DynamicFunctionAdapterContext<T>;
-
-            return _wholeFieldValidationRowDynamic<T>(fnDef, dynamicContext);
-        }
-
-        const functionContext: FunctionAdapterContext<T> = {
-            ...configs,
-            args: configs.args as T,
-            parent: input
-        };
-
-        return _wholeFieldValidationRow(fnDef, functionContext);
-    };
-}
-
-function _wholeFieldValidationRow<T>(
-    fnDef: FieldValidateConfig<T>,
-    configs: FunctionAdapterContext<T>
-) {
-    const results: Record<string, unknown>[] = [];
-    const fn = fnDef.create(configs);
-    const input = configs.parent as Record<string, unknown>[];
-
-    for (let i = 0; i < input.length; i++) {
-        const clone = cloneDeep(input[i]) as Record<string, unknown>;
-
-        if (!isObjectEntity(clone)) {
-            throw new Error(`Invalid record ${JSON.stringify(clone)}, expected an array of simple objects or data-entities`);
-        }
-
-        _processWholeFieldValidationRow(
-            fn,
-            clone,
-            configs.field as string,
-            configs.preserveEmptyObjects,
-            configs.preserveNulls,
-            i,
-            results
-        );
-    }
-
-    return results;
-}
-
-function _wholeFieldValidationRowDynamic<T>(
-    fnDef: FieldValidateConfig<T>,
-    configs: DynamicFunctionAdapterContext<T>,
-) {
-    const results: Record<string, unknown>[] = [];
-    const input = configs.parent as Record<string, unknown>[];
-
-    for (let i = 0; i < input.length; i++) {
-        const clone = cloneDeep(input[i]) as Record<string, unknown>;
-
-        if (!isObjectEntity(clone)) {
-            throw new Error(`Invalid record ${JSON.stringify(clone)}, expected an array of simple objects or data-entities`);
-        }
-
-        const newArgs = configs.args(i, configs.parent) as T;
-        validateFunctionArgs(fnDef, newArgs);
-
-        const fnConfig: FunctionContext<T> = {
-            ...configs,
-            args: newArgs as T,
-        };
-
-        const fn = fnDef.create(fnConfig);
-
-        _processWholeFieldValidationRow(
-            fn,
-            clone,
-            configs.field as string,
-            configs.preserveEmptyObjects,
-            configs.preserveNulls,
-            i,
-            results
-        );
-    }
-
-    return results;
-}
-
-function _processWholeFieldValidationRow(
-    fn: (input: unknown, index: number) => unknown,
-    clone: Record<string, unknown>,
-    field: string,
-    preserveEmptyObjects: boolean,
-    preserveNulls: boolean,
-    i: number,
-    results: unknown[],
-) {
-    const value = get(clone, field);
-    const isValid = fn(value, i);
-    // if it fails validation and we keep null
-    if (!isValid && preserveNulls) {
-        set(clone, field, null);
-        results.push(clone);
-    } else if (!isValid) {
-        // remove key, check if empty record
-        unset(clone, field);
-        // if we preserve empty objects, we don't need to check anything
-        if (preserveEmptyObjects) {
-            results.push(clone);
-        } else {
-            const hasKeys = Object.keys(clone).length !== 0;
-            if (hasKeys) {
-                results.push(clone);
-            }
-        }
-    } else {
-        results.push(clone);
-    }
-}
-
-export function fieldValidationRowExecution<T extends Record<string, any>>(
-    fnDef: FieldValidateConfig<T>,
-    configs: InitialFunctionContext<T>,
-): (input: Record<string, unknown>[]) => Record<string, unknown>[] {
-    return function _fieldValidationRowExecution(
-        input: Record<string, unknown>[],
-    ): Record<string, unknown>[] {
-        if (isNil(configs.field)) throw new Error('Must provide a field option when running a row');
-
-        if (!Array.isArray(input)) {
-            throw new Error('Invalid input, expected an array of objects');
-        }
-
-        if (isFunction(configs.args)) {
-            const dynamicContext = {
-                ...configs,
-                parent: input,
-            } as DynamicFunctionAdapterContext<T>;
-
-            return _fieldValidationRowDynamic<T>(fnDef, dynamicContext);
-        }
-
-        const functionContext: FunctionAdapterContext<T> = {
-            ...configs,
-            args: configs.args as T,
-            parent: input
-        };
-
-        return _fieldValidationRow(fnDef, functionContext);
-    };
-}
-
-function _fieldValidationRow<T>(
-    fnDef: FieldValidateConfig<T>,
-    configs: FunctionAdapterContext<T>
-) {
-    const fn = fnDef.create(configs);
-    const input = configs.parent as unknown[];
-
-    const results: Record<string, unknown>[] = [];
-
-    for (let i = 0; i < input.length; i++) {
-        const clone = cloneDeep(input[i]) as Record<string, unknown>;
-
-        if (!isObjectEntity(clone)) {
-            throw new Error(`Invalid record ${JSON.stringify(clone)}, expected an array of simple objects or data-entities`);
-        }
-
-        _processValidationRow<T>(
-            fn,
-            clone,
-            configs.field as string,
-            configs.preserveEmptyObjects,
-            configs.preserveNulls,
-            i,
-            results
-        );
-    }
-
-    return results;
-}
-
-function _fieldValidationRowDynamic<T>(
-    fnDef: FieldValidateConfig<T>,
-    configs: DynamicFunctionAdapterContext<T>,
-) {
-    const input = configs.parent as unknown[];
-
-    const results: Record<string, unknown>[] = [];
-
-    for (let i = 0; i < input.length; i++) {
-        const newArgs = configs.args(i, configs.parent) as T;
-        validateFunctionArgs(fnDef, newArgs);
-
-        const fnConfig: FunctionContext<T> = {
-            ...configs,
-            args: newArgs as T,
-        };
-
-        const fn = fnDef.create(fnConfig);
-        const clone = cloneDeep(input[i]) as Record<string, unknown>;
-
-        if (!isObjectEntity(clone)) {
-            throw new Error(`Invalid record ${JSON.stringify(clone)}, expected an array of simple objects or data-entities`);
-        }
-
-        _processValidationRow<T>(
-            fn,
-            clone,
-            configs.field as string,
-            configs.preserveEmptyObjects,
-            configs.preserveNulls,
-            i,
-            results
-        );
-    }
-
-    return results;
-}
-
-function _processValidationRow<T>(
-    fn: (input: unknown, index: number) => unknown,
-    clone: Record<string, unknown>,
-    field: string,
-    preserveEmptyObjects: boolean,
-    preserveNulls: boolean,
-    i: number,
-    results: unknown[],
-) {
-    const value: unknown = get(clone, field);
-
-    if (Array.isArray(value)) {
-        const fieldList = callValue(fn, value, preserveNulls, true, i);
-
-        // we have results in list or we don't care if its an empty list here
-        if (fieldList.length > 0) {
-            set(clone, field, fieldList);
-            results.push(clone);
-        } else {
-            unset(clone, field);
-
-            if (preserveEmptyObjects) {
-                results.push(clone);
-            } else {
-                const hasKeys = Object.keys(clone).length !== 0;
-                if (hasKeys) {
-                    results.push(clone);
-                }
-            }
-        }
-    } else {
-        const isValid = fn(value as T, i);
-        // if it fails validation and we keep null
-        if (!isValid && preserveNulls) {
-            set(clone, field, null);
-            results.push(clone);
-        } else if (!isValid) {
-            // remove key, check if empty record
-            unset(clone, field);
-            // if we preserve empty objects, we don't need to check anything
-            if (preserveEmptyObjects) {
-                results.push(clone);
-            } else {
-                const hasKeys = Object.keys(clone).length !== 0;
-                if (hasKeys) {
-                    results.push(clone);
-                }
-            }
-        } else {
-            results.push(clone);
-        }
-    }
-}
