+++ packages/data-mate/src/adapters/function-adapter/validators/field/index.ts
@@ -1,0 +1,4 @@
+export * from './fieldColumn';
+export * from './fieldRow';
+export * from './wholeFieldColumn';
+export * from './wholeFieldRow';
